moved multi-agent file in the correct folder
